+++ Tests/Web.Tests/V1/DevicesControllerTests.cs
@@ -8,19 +8,19 @@ using Domain.Shared;
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
-using Presentation.Tests.Fixtures;
+using Presentation.Tests.V1.Fixtures;
 using Xunit;
 
-namespace Presentation.Tests;
+namespace Presentation.Tests.V1;
 
 public class DevicesControllerTests : IClassFixture<DevicesControllerFixture>
 {
     private readonly DevicesControllerFixture _fixture;
 
-	public DevicesControllerTests(DevicesControllerFixture fixture)
-	{
-		_fixture = fixture;
-	}
+    public DevicesControllerTests(DevicesControllerFixture fixture)
+    {
+        _fixture = fixture;
+    }
 
     [Fact]
     public async Task GetAsync_Should_ReturnActionResultOfPaginatedModelOfGetDeviceResponse_WhenDataIsValid()
