+++ Web/Configurations/ApiVersioningConfiguration.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Versioning;
+
+namespace Web.Configurations;
+
+public static class ApiVersioningConfiguration
+{
+    public static void ConfigureApiVersioning(this IServiceCollection services)
+    {
+        services
+            .AddApiVersioning(options =>
+            {
+                options.DefaultApiVersion = new ApiVersion(1, 0);
+                options.AssumeDefaultVersionWhenUnspecified = true;
+                options.ReportApiVersions = true;
+                //options.ApiVersionReader = ApiVersionReader.Combine(
+                //    new QueryStringApiVersionReader(),
+                //    new HeaderApiVersionReader(),
+                //    new MediaTypeApiVersionReader());
+                options.ApiVersionReader = new UrlSegmentApiVersionReader();
+            })
+            .AddVersionedApiExplorer(options =>
+            {
+                options.GroupNameFormat = "'v'VVV";
+                options.SubstituteApiVersionInUrl = true;
+            });
+    }
+}
