+++ Web/Configurations/AutoMapperConfiguration.cs
@@ -1,0 +1,9 @@
+﻿namespace Web.Configurations;
+
+public static class AutoMapperConfiguration
+{
+    public static void ConfigureAutoMapper(this IServiceCollection services)
+    {
+        services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
+    }
+}
